Update hiring process instructions for new job data file
Rather than moving the instructions from the README directly into the
handbook, we link to the README from the handbook.

The thinking is that the instructions are fairly technical and detailed,
and will likely only need to be read once to "get it", and would
otherwise permanently clutter up the Hiring Handbook.