Fix various grammar issues.
"fewer" is preferred over "less" when referring to discrete nouns (https://en.wikipedia.org/wiki/Fewer_vs._less).

In "ask others for help and advise", I chose to interpret that as "ask others for help and advice" rather than a request to both "ask for help" and "advise".

Use "who" to refer to people instead of "that" (http://english.stackexchange.com/a/428).

Match the tense of "is constrained by" in the following "we _have_ added", since both are ongoing.