Update production engineer role to include go development
Also include the kind of projects that we are working on as I recently
heard that it is way much more inclusive to explain what kind of work we
do in a role than to set the checklist of requirements.

The idea is that people get more excited about the opportunities
presented than about what bar they have to pass to be even considered.