I2P demo: Restore using non-root user account
Turns out the installation as-is on the I2P demo doesn't require the external confirmation. Restore the using a non-root user so that Gravatar is present, and normalizes the workflow for an individual user. Showing that you don't have to be GitLab admin to do I2P end to end.